Move vendor search to it's own route
