+++ tests/pages/search-vendors.js
